seeded database with nfl info

+++ db/migrate/20171009174906_create_nfl_athletes.rb
@@ -10,9 +10,9 @@ class CreateNflAthletes < ActiveRecord::Migration[5.1]
     	t.integer :rushing_touchdowns, :default => 0
     	t.integer :receiving_yards, :default => 0
     	t.integer :receiving_touchdowns, :default => 0
-    	t.integer :stock_value, :default => 100
-    	t.integer :fantasy_projection, :default => 0
-    	t.integer :actual_fantasy_value, :default => 0
+    	t.float :stock_value, :default => 100
+    	t.float :fantasy_projection, :default => 0
+    	t.float :actual_fantasy_value, :default => 0
     	t.timestamps
     end
   end
